+++ makedocs.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+. ~/venv/checkdmarc/bin/activate
+cd docs && make html && cp -r build/html/* ../../checkdmarc-docs/